Move all util classes into util directory
