+++ src/core/utils/StringUtils.ts
